# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type AdminLoginResultData = {
  token: string;
  username: string;
};

export interface AdminLoginResult {
  success: boolean;
  data: AdminLoginResultData;
}

export interface AdminStats {
  total: number;
  pending: number;
  approved: number;
  rejected: number;
}

export interface AdminStatsResponse {
  success: boolean;
  data: AdminStats;
}

export interface DeviceHandshakeInput {
  /** @minLength 1 */
  device_id: string;
  /**
     * Stable identifier for the application using DeviceGuard
     * @minLength 1
     */
  app_id: string;
  app_name?: string;
  package_name?: string;
  device_name?: string;
  android_version?: string;
  manufacturer?: string;
  model?: string;
  sdk_version?: string;
  app_version?: string;
  requested_permissions?: string[];
}

export interface LicenseValidationInput {
  /** @minLength 1 */
  license_token: string;
}

export type DeviceHandshakeResultStatus = typeof DeviceHandshakeResultStatus[keyof typeof DeviceHandshakeResultStatus];


export const DeviceHandshakeResultStatus = {
  approved: 'approved',
  pending: 'pending',
  rejected: 'rejected',
  expired: 'expired',
  error: 'error',
} as const;

export interface DeviceHandshakeResult {
  access: boolean;
  status: DeviceHandshakeResultStatus;
  message: string;
  device_id?: string;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  license_token?: string | null;
  granted_permissions?: string[];
  missing_permissions?: string[];
}

export type DeviceStatus = typeof DeviceStatus[keyof typeof DeviceStatus];


export const DeviceStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  BANNED: 'BANNED',
  EXPIRED: 'EXPIRED',
} as const;

export interface Device {
  id: string;
  device_id: string;
  device_id_hash: string;
  app_id: string;
  /** @nullable */
  app_name?: string | null;
  /** @nullable */
  package_name?: string | null;
  /** @nullable */
  device_name?: string | null;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  android_version?: string | null;
  /** @nullable */
  sdk_version?: string | null;
  /** @nullable */
  app_version?: string | null;
  /** @nullable */
  ip_address?: string | null;
  requested_permissions?: string[];
  granted_permissions?: string[];
  status: DeviceStatus;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  notes?: string | null;
  registered_at: string;
  last_active_at: string;
}

export type DevicesResponseData = {
  devices: Device[];
};

export interface DevicesResponse {
  success: boolean;
  data: DevicesResponseData;
}

export interface ApproveDeviceInput {
  /** @minimum 0 */
  duration_days: number;
  granted_permissions: string[];
}

export interface ActionResponse {
  success: boolean;
  message: string;
}

export interface ActivityLog {
  id: string;
  /** @nullable */
  device_id?: string | null;
  /** @nullable */
  device_name?: string | null;
  action: string;
  /** @nullable */
  ip_address?: string | null;
  created_at: string;
}

export type LogsResponseData = {
  logs: ActivityLog[];
};

export interface LogsResponse {
  success: boolean;
  data: LogsResponseData;
}

export type ListAdminDevicesParams = {
status?: ListAdminDevicesStatus;
search?: string;
};

export type ListAdminDevicesStatus = typeof ListAdminDevicesStatus[keyof typeof ListAdminDevicesStatus];


export const ListAdminDevicesStatus = {
  ALL: 'ALL',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  BANNED: 'BANNED',
  EXPIRED: 'EXPIRED',
} as const;