# lib/api-zod/src/generated/types/actionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActionResponse {
  success: boolean;
  message: string;
}