# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: string;
  /** @nullable */
  device_id?: string | null;
  /** @nullable */
  device_name?: string | null;
  action: string;
  /** @nullable */
  ip_address?: string | null;
  created_at: string;
}