# lib/api-zod/src/generated/types/adminStatsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStats } from './adminStats';

export interface AdminStatsResponse {
  success: boolean;
  data: AdminStats;
}