# lib/api-zod/src/generated/types/deviceHandshakeResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceHandshakeResultStatus } from './deviceHandshakeResultStatus';

export interface DeviceHandshakeResult {
  access: boolean;
  status: DeviceHandshakeResultStatus;
  message: string;
  device_id?: string;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  license_token?: string | null;
  granted_permissions?: string[];
  missing_permissions?: string[];
}