# lib/api-zod/src/generated/types/devicesResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DevicesResponseData } from './devicesResponseData';

export interface DevicesResponse {
  success: boolean;
  data: DevicesResponseData;
}