# lib/api-zod/src/generated/types/logsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogsResponseData } from './logsResponseData';

export interface LogsResponse {
  success: boolean;
  data: LogsResponseData;
}